requirements.txt
langchain
ipykernel
python-dotenv
langchain-community
pypdf
bs4
langchain-text-splitters
sentence-transformers
chromadb
langchain-huggingface
faiss-cpu
streamlit
langchain_ollama
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette